requirements.txt
# MINIMAL GOOGLE-ONLY TRAVEL AI DEMO
# Only essential dependencies for universal travel demo

# Google AI services (PRIMARY REQUIREMENT)
google-generativeai>=0.3.0

# Environment configuration
python-dotenv>=1.0.0

# OPTIONAL: Full backend API (comment out if only using universal_travel_demo.py)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
